Create week.c
a program that retrieves any valid date and calculates the date one week from now.

Expect to see multiple date lines in the input in standard Slovak format. The last line is blank. Load all the dates and if the date is valid then calculate the date of the week to it. If the date is invalid, then print an error message instead and continue with the next date.

The date on the input is made up of 3 numbers separated by a period. The first number is the day number, the second number indicates the month, and the third indicates the year. The date is valid if it exists in the Slovak calendar. If the year is leap year, then February 29 is valid in it.

Dates before 1900 and after 2100 are considered invalid.